Refactor with hexagonal architecture
