+++ Adapters/routes/skeleton.js
