+++ infraestructure/index.js
@@ -2,4 +2,4 @@
 // De lo contrario: Cambiar la estructura interna de los archivos o la firma puede romper la interfaz con los clientes.
 
 module.exports.mongoDbConfig = require("./mongodb.config.js");
-module.exports.mongoMigrationConfig = require("./mongo.migration.config.js");
+module.exports.mongoMigrationConfig = require("./mongo.migration.config.js.js");
