+++ tests/server.js
@@ -5,7 +5,7 @@ const express = require("express");
 const cors = require("cors");
 
 const app = express();
-const routes = require("../routes");
+const routes = require("../Adapters/routes");
 
 app.use(express.json());
 app.use(cors());
